--- lib/crichton-representors.rb
@@ -1,8 +1,0 @@
-require 'crichton/options'
-require 'crichton/field'
-require 'crichton/transition'
-require 'crichton/representor'
-require 'crichton/representors/representor_builder'
-require 'crichton/representors/representor_hash'
-require 'crichton/deserializers/hal_deserializer'
-require 'crichton/deserializers/deserialization_error'
