+++ lib/representors.rb
@@ -1,0 +1,8 @@
+require 'representors/options'
+require 'representors/field'
+require 'representors/transition'
+require 'representors/representor'
+require 'representors/representor_builder'
+require 'representors/representor_hash'
+require 'representors/deserializers/hal_deserializer'
+require 'representors/deserializers/deserialization_error'
