+++ lib/representors/version.rb
@@ -1,0 +1,4 @@
+# Used to prevent the class/module from being loaded more than once
+module Representors
+  VERSION = '0.0.1'
+end
