+++ representors.gemspec
@@ -1,12 +1,12 @@
 $LOAD_PATH.unshift('lib')
-require 'crichton/representors/version'
+require 'representors/version'
 
 Gem::Specification.new do |s|
-  s.name = 'crichton-representors'
-  s.version = Crichton::Representor::VERSION::STRING
+  s.name = 'representors'
+  s.version = Representors::VERSION
   s.date = Time.now.strftime('%Y-%m-%d')
   s.summary = 'It has the knowledge of Hypermedia media-types from the Ancients!'
-  s.homepage = 'http://github.com//crichton-representors'
+  s.homepage = 'http://github.com/representors'
   s.email = ''
   s.authors = ['Mark W. Foster', 'Shea Valentine']
   s.files = ['lib/**/*', 'spec/**/*', 'tasks/**/*', '[A-Z]*'].map { |glob| Dir[glob] }.inject([], &:+)
@@ -14,7 +14,9 @@ Gem::Specification.new do |s|
   s.rdoc_options = ['--main', 'README.md']
 
   s.description = <<-DESC
-    Crichton Representors is a library to simplify Hypermedia message representation.
+    Crichton Representors is a library containing serializers and deserializers to and from hypermedia formats.
+    This library does not have the functionality to get and post data over the Internet. Consider Farscape for that.
+    This library also does not automatically decorates objects. Consider Crichton for that.
   DESC
 
   s.add_dependency('enumerable-lazy', '~> 0.0.1') if RUBY_VERSION < '2.0'
